[cartesians] NBList uses another criterion to trigger update
Now the NBList keeps track of an accumulated displacement for each atom and checks whether the atom has travelled further than the allowed cutoff. If so, an update for that atom is triggered.

The previous implementation required evaluation of a distance between old and new positions. This change however has little impact on the overall simulation time, as the update is still done at O(N) cost.