[surpass] Random generation of a starting conformation; PDB output
The system uses bioshell-pdb::nerf module to build a random chain conformation. The constructor randomly selects planar and dihedraln angles; restore_linear_chain() is used to restore Cartesian coordinates.

A conformation may be writen in the PDB format; this is just a temporary solution that should be fixed after bioshell-pdb::PdbAtom::to_pdb() is implemented.