[surpass] Fixing bugs in movers
Movers were breaking a chain when pivot atoms vere located on either side of a periodic box boundary. The fix required introduction of  SurpassAlphaSystem::ca_to_nearest_vec3() the get the nearest image of an atom. Also a bunch of changes in TailMove::compute_move() and HingeMove::compute_move() were made

A new test in test_movers.rs reads a PDB file with a pose and makes the move that preiously failed.

The check_bond_lengths() function in surpass.rs makes a detailed check if bond lengths are preserved. If so, two PDB files are writen: with the conformation before and after the failing move.