[pdb] phi() and psi() evaluating functions.
This update required also implementing a way to get the next and the previous residue that follows / precedes a given ResidueId struct. Currently this is solved by storing all the ResidueId for every residue in a vec. ResidueId struct implements now Ord and PartialOrd traits, so the Vec<ResidueId> cen be binary-searched to find the index of the residue; given that index one can take previous or next ResidueId (just by +1 / -1 to that index)