[bioshell-cif] CifTable can create a table of a single row directly from fields.
This change unifies parsing CIF files, there a collection of elements (say, a list of alpha-helices) is provided in a loop-block while a single element of the same type (just one helix) as data items.

Also in this commit: the list of keywords is parsed from a CIF file and stored in the Structure struct.

Unit tests passed